Fix su eliminazione cartella installazione pacchetti se questa non esiste

+++ Artifacts/windows-atlas-postazione-studente/install-choco-packages.ps1
@@ -116,7 +116,12 @@ function Install-Packages
         }
 
 		$packageDestPath = "$HOME/Desktop/InstallDir/"
-		Remove-Item -LiteralPath $packageDestPath -Force -Recurse
+		$packageDestPathExists = Test-Path $packageDestPath -PathType Container
+		if ($packageDestPathExists)
+		{
+			Remove-Item -LiteralPath $packageDestPath -Force -Recurse
+		}
+		
 		md $packageDestPath
 		$nupkgurl = "https://atlas-testing.webscience.it/artifacts/AtlasPostazioneStudente/atlaspostazionestudente.3.0.0.nupkg"
 		$nuspecurl= "https://atlas-testing.webscience.it/artifacts/AtlasPostazioneStudente/atlaspostazionestudente.nuspec"
